extract types extract components
