+++ src/app/designs/page.tsx
@@ -1,14 +1,10 @@
 "use client";
 
-import { Box, Typography, IconButton } from "@mui/material";
+import { Box } from "@mui/material";
 import Link from "next/link";
 import { useRouter } from "next/navigation";
 import { useEffect, useState, useCallback, useMemo } from "react";
-import HomeIcon from "@mui/icons-material/Home";
 import Footer from "../../components/layout/Footer";
-import Image, { StaticImageData } from "next/image";
-import ArrowCircleLeftIcon from "@mui/icons-material/ArrowCircleLeft";
-import ArrowCircleRightIcon from "@mui/icons-material/ArrowCircleRight";
 import {
     figmaDesigns,
     figmaImages,
@@ -16,274 +12,9 @@ import {
     HELIKA_UA_IMAGES,
 } from "@/constants/constants";
 import React from "react";
-
-// Types
-interface CarouselProps {
-    images: (string | StaticImageData)[];
-    currentIndex: number;
-    onIndexChange: (index: number) => void;
-    title: string;
-    textColor: string;
-}
-
-interface ImageCarouselProps {
-    images: (string | StaticImageData)[];
-    currentIndex: number;
-    onIndexChange: (index: number) => void;
-}
-
-interface HomeButtonProps {
-    onClick: () => void;
-    textColor: string;
-    bgColor: string;
-}
-
-interface ProgressBarProps {
-    progressKey: number;
-}
-
-// Reusable components
-const HomeButton = ({ onClick, textColor, bgColor }: HomeButtonProps): React.ReactElement => (
-    <button
-        onClick={onClick}
-        style={{
-            position: "absolute",
-            top: "20px",
-            left: "20px",
-            padding: "10px",
-            width: "50px",
-            height: "50px",
-            borderRadius: "50%",
-            border: "1px solid",
-            background: textColor,
-            color: bgColor,
-            cursor: "pointer",
-            transition: "all 0.3s ease",
-            display: "flex",
-            justifyContent: "center",
-            alignItems: "center",
-            borderColor: bgColor,
-        }}
-    >
-        <HomeIcon
-            style={{
-                fontSize: "24px",
-                height: "24px",
-                width: "24px",
-                color: bgColor,
-            }}
-        />
-    </button>
-);
-
-const ProgressBar = ({ progressKey }: ProgressBarProps): React.ReactElement => (
-    <Box
-        sx={{
-            position: "absolute",
-            bottom: 0,
-            left: 0,
-            width: "100%",
-            height: "5px",
-            backgroundColor: "rgba(255, 255, 255, 0.3)",
-            zIndex: 200,
-            overflow: "hidden",
-        }}
-    >
-        <Box
-            sx={{
-                height: "100%",
-                backgroundColor: "white",
-                animation: "fillProgress 5s linear infinite",
-                width: "100%",
-                zIndex: 201,
-            }}
-            key={progressKey}
-        />
-        <style jsx>{`
-            @keyframes fillProgress {
-                from {
-                    width: 0;
-                }
-                to {
-                    width: 100%;
-                }
-            }
-        `}</style>
-    </Box>
-);
-
-const ImageCarousel = ({ images, currentIndex, onIndexChange }: ImageCarouselProps): React.ReactElement => {
-    const maxVisible = 3;
-    const carouselIndex = useMemo(
-        () => Math.min(currentIndex, images.length - maxVisible),
-        [currentIndex, images.length]
-    );
-
-    return (
-        <Box
-            sx={{
-                height: { xs: "15vh", sm: "20vh" },
-                justifyContent: "center",
-                alignItems: "center",
-                marginTop: { xs: "8px", sm: "16px" },
-                display: "flex",
-                flexDirection: "row",
-                gap: { xs: "0.5em", sm: "1em" },
-                position: "relative",
-                overflow: "hidden",
-            }}
-        >
-            <IconButton
-                onClick={() => onIndexChange(Math.max(currentIndex - 1, 0))}
-                sx={{
-                    position: "absolute",
-                    left: { xs: "4px", sm: "0" },
-                    top: "50%",
-                    transform: "translateY(-50%)",
-                    zIndex: 10,
-                    background: "rgba(0, 0, 0, 0.5)",
-                    color: "white",
-                    "&:hover": { background: "rgba(0, 0, 0, 0.7)" },
-                    padding: { xs: "4px", sm: "8px" },
-                    minWidth: { xs: "32px", sm: "48px" },
-                    minHeight: { xs: "32px", sm: "48px" },
-                }}
-            >
-                <ArrowCircleLeftIcon sx={{ fontSize: { xs: "1.5em", sm: "3em" } }} />
-            </IconButton>
-            <Box
-                sx={{
-                    display: "flex",
-                    flexDirection: "row",
-                    gap: { xs: "0.5em", sm: "1em" },
-                    transform: `translateX(-${carouselIndex * 33.33}%)`,
-                    transition: "transform 0.3s ease",
-                    width: `${images.length * 33.33}%`,
-                    paddingLeft: { xs: "2em", sm: "1em" },
-                    paddingRight: { xs: "2em", sm: "1em" },
-                }}
-            >
-                {images.map((image, idx) => (
-                    <Box
-                        key={idx}
-                        sx={{
-                            height: "100%",
-                            aspectRatio: "3593/2090",
-                            backgroundColor: "white",
-                            overflow: "hidden",
-                            cursor: "pointer",
-                            display: "flex",
-                            justifyContent: "center",
-                            alignItems: "center",
-                            flex: "0 0 calc(33.33% - 0.5em)",
-                        }}
-                        onMouseEnter={() => onIndexChange(idx)}
-                    >
-                        <Image
-                            src={image}
-                            alt={`Thumbnail ${idx + 1}`}
-                            style={{
-                                width: "100%",
-                                height: "100%",
-                                objectFit: "cover",
-                            }}
-                            loading="lazy"
-                        />
-                    </Box>
-                ))}
-            </Box>
-            <IconButton
-                onClick={() =>
-                    onIndexChange(
-                        Math.min(currentIndex + 1, images.length - maxVisible)
-                    )
-                }
-                sx={{
-                    position: "absolute",
-                    right: { xs: "4px", sm: "0" },
-                    top: "50%",
-                    transform: "translateY(-50%)",
-                    zIndex: 10,
-                    background: "rgba(0, 0, 0, 0.5)",
-                    color: "white",
-                    "&:hover": { background: "rgba(0, 0, 0, 0.7)" },
-                    padding: { xs: "4px", sm: "8px" },
-                    minWidth: { xs: "32px", sm: "48px" },
-                    minHeight: { xs: "32px", sm: "48px" },
-                }}
-            >
-                <ArrowCircleRightIcon sx={{ fontSize: { xs: "1.5em", sm: "3em" } }} />
-            </IconButton>
-        </Box>
-    );
-};
-
-const CarouselSection = ({ images, currentIndex, onIndexChange, title, textColor }: CarouselProps): React.ReactElement => (
-    <Box
-        sx={{
-            height: { xs: "auto", sm: `calc(100vh + min(25vh, 25vw))` },
-            width: "100%",
-            display: "flex",
-            flexDirection: "column",
-            alignItems: "center",
-            scrollSnapAlign: "start",
-            color: textColor,
-            justifyContent: "start",
-            paddingTop: { xs: "2em", sm: "5em" },
-            paddingBottom: { xs: "2em", sm: "0" },
-        }}
-    >
-        <Typography
-            variant="h3"
-            sx={{
-                color: textColor,
-                fontFamily: "inherit",
-                paddingBottom: { xs: "0.25em", sm: "0.5em" },
-                fontSize: { xs: "1.5rem", sm: "2.5rem" },
-                textAlign: "center",
-            }}
-        >
-            {title}
-        </Typography>
-        <Box
-            id="designsbox"
-            sx={{
-                height: { xs: "40vh", sm: "60vh" },
-                width: { xs: "90%", sm: "auto" },
-                maxWidth: { xs: "100%", sm: "90vw" },
-                aspectRatio: "3593/2090",
-                position: "relative",
-                boxShadow: "10px 0 10px -4px gray, -10px 0 10px -4px gray",
-                cursor: "pointer",
-                margin: "0 auto",
-                display: "flex",
-                justifyContent: "center",
-                alignItems: "center",
-                overflow: "hidden",
-            }}
-        >
-            <Image
-                src={images[currentIndex]}
-                alt={`${title} image ${currentIndex + 1}`}
-                style={{
-                    objectFit: "contain",
-                    zIndex: 100,
-                    marginBottom: "1em",
-                    width: "100%",
-                    height: "100%",
-                }}
-                fill={true}
-                priority={currentIndex === 0}
-            />
-            <ProgressBar progressKey={currentIndex} />
-        </Box>
-        <ImageCarousel
-            images={images}
-            currentIndex={currentIndex}
-            onIndexChange={onIndexChange}
-        />
-    </Box>
-);
+import HomeButton from "@/components/common/HomeButton";
+import CarouselSection from "@/components/common/CarouselSection";
+import Image from "next/image";
 
 export default function Designs(): React.ReactElement {
     const router = useRouter();
