+++ src/types/about.ts
@@ -1,0 +1,7 @@
+import { ElementType } from "react";
+
+export interface SocialLinkProps {
+    href: string;
+    icon: ElementType;
+    text: string;
+}
