+++ src/types/gallery.ts
@@ -1,0 +1,6 @@
+export interface ImageData {
+    src: string;
+    alt: string;
+    width: number;
+    height: number;
+}
